feat: add delete confirmation dialog for list items

+++ lib/src/features/presentation/screens/wish_list/wish_list_details/wish_list_line/wish_list_line_widget.dart
@@ -200,9 +200,18 @@ class WishListLineWidget extends StatelessWidget {
         if (isDeleteButtonVisible)
           InkWell(
             onTap: () async {
-              await context.read<WishListDetailsCubit>().deleteWishListLine(
-                    wishListLineEntity,
-                  );
+              confirmDialog(
+                context: context,
+                title: LocalizationConstants.deleteListItem.localized(),
+                message: LocalizationConstants.areYouSureYouWantToDeleteThisItem
+                    .localized(),
+                confirmText: LocalizationConstants.delete.localized(),
+                onConfirm: () async {
+                  await context.read<WishListDetailsCubit>().deleteWishListLine(
+                        wishListLineEntity,
+                      );
+                },
+              );
             },
             child: Padding(
               padding: const EdgeInsets.all(10),
